+++ src/api/create-layer.ts
@@ -1,5 +1,6 @@
 import { computed, type Component } from 'vue'
 import type {
+  AnyComponent,
   LayerConfigCreateOf,
   LayerConfigContentOf,
   LayerConfigFragmentCreate,
@@ -15,11 +16,9 @@ import {
 import { createLayerInstance } from '@/runtime/layer-instance'
 import { toValue, type MaybeRefOrGetter } from '@/compat'
 
-export function createLayer<CContainer extends Component>(
+export function createLayer<CContainer extends AnyComponent>(
   Container: CContainer,
-  config: MaybeRefOrGetter<
-    LayerConfigCreateOf<PropsOf<CContainer>>
-  > = {},
+  config: MaybeRefOrGetter<LayerConfigCreateOf<PropsOf<CContainer>>> = {},
 ) {
   type ContainerP = PropsOf<CContainer>
 
@@ -28,7 +27,7 @@ export function createLayer<CContainer extends Component>(
     return {
       ...mergeFragment(
         toFragmentFromContainer(containerConfig),
-        { container: { component: Container } },
+        { container: { component: Container as Component } },
       ),
       ...(adapter !== undefined ? { adapter } : {}),
     }
@@ -38,23 +37,28 @@ export function createLayer<CContainer extends Component>(
     Content?: undefined,
     useConfig?: MaybeRefOrGetter<LayerConfigContentOf<LooseProps, ContainerP>>,
   ): LayerInstance<LooseProps, ContainerP>
-  function useLayer<CContent extends Component>(
+  function useLayer<CContent extends AnyComponent>(
     Content: CContent,
     useConfig?: MaybeRefOrGetter<
       LayerConfigContentOf<PropsOf<CContent>, ContainerP>
     >,
   ): LayerInstance<PropsOf<CContent>, ContainerP>
   function useLayer(
-    Content?: Component,
+    Content?: AnyComponent,
     useConfig: MaybeRefOrGetter<LayerConfigContentOf<any, ContainerP>> = {},
   ): LayerInstance<any, ContainerP> {
     const use = computed(() =>
       mergeFragment(
         toFragmentFromContent(toValue(useConfig)),
-        Content ? { content: { component: Content } } : undefined,
+        Content
+          ? { content: { component: Content as Component } }
+          : undefined,
       ),
     )
-    return createLayerInstance({ create, use }) as LayerInstance<any, ContainerP>
+    return createLayerInstance({ create, use }) as LayerInstance<
+      any,
+      ContainerP
+    >
   }
 
   return useLayer